requirement.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Dance==7.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
flask-redis==0.4.0
Flask-Security==5.6.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
h11==0.16.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
passlib==1.7.4
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
qrcode==8.1
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
URLObject==2.4.3
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
